fix bugs and can run but no lock for ssd_hdr and ssd_table

+++ src/ssd_buf_table.h
@@ -1,9 +1,9 @@
 #ifndef SSDBUFTABLE_H
 #define SSDBUFTABLE_H
 
-extern void initSSDBufTable(int size);
-extern unsigned ssdbuftableHashcode(SSDBufferTag *ssd_buf_tag);
-extern int ssdbuftableLookup(SSDBufferTag *ssd_buf_tag, unsigned hash_code);
-extern int ssdbuftableInsert(SSDBufferTag *ssd_buf_tag, unsigned hash_code, int ssd_buf_id);
-extern int ssdbuftableDelete(SSDBufferTag *ssd_buf_tag, unsigned hash_code);
+extern void initSSDBufTable(size_t size);
+extern unsigned long ssdbuftableHashcode(SSDBufferTag *ssd_buf_tag);
+extern long ssdbuftableLookup(SSDBufferTag *ssd_buf_tag, unsigned long hash_code);
+extern long ssdbuftableInsert(SSDBufferTag *ssd_buf_tag, unsigned long hash_code, size_t ssd_buf_id);
+extern long ssdbuftableDelete(SSDBufferTag *ssd_buf_tag, unsigned long hash_code);
 #endif   /* SSDBUFTABLE_H */
